document train and test
