+++ openproblems/tasks/denoising/README.md
@@ -7,6 +7,6 @@ A key challenge in evaluating denoising methods is the general lack of a ground 
 
 ## API
 
-Datasets only need to contain the raw UMI counts (in `adata.X`).
+Datasets should contain the raw UMI counts in `adata.X`, subsampled to training (`adata.obsm["train"]`) and testing (`adata.obsm["test"]`) datasets using `openproblems.tasks.denoising.datasets.utils.split_data`.
 
 The task-specific data loader functions should split the provided raw UMI counts into a training and a test dataset, as described by <a href="https://www.biorxiv.org/content/10.1101/786269v1">Batson et al., 2019</a>. The training dataset should be stored in `adata.obsm['train']`, and the test dataset should be stored in `adata.obsm['test']`. Methods should store the denoising result in `adata.obsm['denoised']`.
